Moved `getAccessibilityRepresentation` inside `BaseBancomatCard`
